Rename pact to driver
