+++ src/PhpPact/Consumer/Driver/DriverInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpPact\Consumer\Driver;
+
+interface DriverInterface
+{
+}
